requirements.txt
Brotli==1.0.9
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
ConfigArgParse==1.5.3
Faker==13.3.4
Flask==2.1.1
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==21.12.0
geventhttpclient==1.5.3
google-api-core==2.7.1
google-auth==2.6.2
google-cloud-bigquery==3.0.1
google-cloud-bigquery-storage==2.13.0
google-cloud-core==2.2.3
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.45.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.0.3
locust==2.8.5
MarkupSafe==2.1.1
msgpack==1.0.3
numpy==1.22.3
packaging==21.3
proto-plus==1.20.3
protobuf==3.20.0
psutil==5.9.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.27.1
roundrobin==0.0.2
rsa==4.8
six==1.16.0
typing-extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
